Save statistics client side too.

+++ core/src/main/java/ee/taltech/voshooter/networking/messages/clientreceived/GameEnd.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class GameEnd {
 
     public int gameMode;
-    public int gameLength;
+    public double gameLength;
     public int playerCount;
     public int botAmount;
     public GameMap.MapType mapType;
@@ -17,7 +17,7 @@ public class GameEnd {
     public GameEnd() {
     }
 
-    public GameEnd(int gameMode, int gameLength, int playerCount, int botAmount, GameMap.MapType mapType,
+    public GameEnd(int gameMode, double gameLength, int playerCount, int botAmount, GameMap.MapType mapType,
                    List<String> leaderBoard) {
         this.gameMode = gameMode;
         this.gameLength = gameLength;
